ML/requirements.txt
Flask
flask-cors
selenium
webdriver-manager
beautifulsoup4
python-dotenv
google-generativeai
googletrans==4.0.0-rc1
